testing/web-platform/tests/websockets/interfaces/CloseEvent/002.html

<!doctype html>
<title>WebSockets: dispatching a CloseEvent</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../constants.js?pipe=sub></script>
<div id=log></div>
<script>
async_test(function(t) {
  var e = document.createEvent('CloseEvent');
  e.initCloseEvent('foo', false, false, false);
  var div = document.createElement('div');
  div.addEventListener('foo', t.step_func(function(ev) {
    assert_equals(ev.bubbles, false);
    assert_equals(ev.cancelable, false);
    assert_equals(ev.wasClean, false);
    t.done();
  }), false);
  div.dispatchEvent(e);
});
</script>
